+++ TPie/Plugin.cs
@@ -95,7 +95,7 @@ namespace TPie
                 AssemblyLocation = Assembly.GetExecutingAssembly().Location;
             }
 
-            Version = Assembly.GetExecutingAssembly().GetName().Version?.ToString() ?? "1.8.1.0";
+            Version = Assembly.GetExecutingAssembly().GetName().Version?.ToString() ?? "1.8.2.0";
 
             Framework.Update += Update;
             UiBuilder.Draw += Draw;